Move and rename SceneUtil
